requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
datallm==0.2.6
Faker==30.6.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
mostlyai==0.4.3
numpy==2.1.2
pandas==2.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydotenv==0.0.7
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rich==13.9.2
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
typing_extensions==4.12.2
tzdata==2024.2
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
datallm==0.2.6
Faker==30.6.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
mostlyai==0.4.3
numpy==2.1.2
pandas==2.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydotenv==0.0.7
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rich==13.9.2
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
typing_extensions==4.12.2
tzdata==2024.2
